clippy: five lint classes down, and the measurement trap I fell into
State now, with the WHOLE workspace compiling: 0 errors, 14619 clippy warnings.

CORRECTION TO THE PREVIOUS COMMIT MESSAGE

It claimed 17134 -> 10059. That 10059 was measured while rustre-debug and its
dependents did not compile, and a crate that fails to compile emits no lints —
so the total dropped without anything improving. I described that exact trap in
the same commit and then fell into it two paragraphs later. The real figure with
everything building is 14619.

Totals in this repo move with build state and are therefore not comparable
across runs while other sessions edit concurrently. PER-CLASS counts are:

  must_use_candidate      1131 -> 313
  missing_const_for_fn    1291 -> 476
  cast_lossless            526 ->  81
  uninlined_format_args    195 ->  55
  doc_markdown             431 -> 367

Those reductions are a LOWER bound, not an estimate: the baseline was
under-counted for the same reason, so the true starting numbers were higher.
The classes that appear to have grown (module_name_repetitions 974 -> 1372,
match_same_arms 503 -> 666, missing_errors_doc 628 -> 846) are the same artifact
seen from the other side — those crates were missing from the baseline count.

UNBLOCKED THE BUILD

`rustre-debug/src/ios/unwind.rs` had EIGHT byte-identical copies of
`a_frameless_leaf_with_a_zero_sized_frame_still_has_a_caller` — verified
identical by hashing each block, one digest for all eight. Seven duplicates
removed. That crate did not compile, which blocked its own session as much as
this one, and made every workspace-wide number meaningless while it lasted.

Sample suites after the fixes: arch-68k 390, arch-z80 490, loader-pdf 352,
trace 356, diff-bindiff 361 — all green, 0 failed.

+++ crates/rustre-arch-sparc/src/sparc_emulator.rs
@@ -310,7 +310,7 @@ impl SparcState {
         let neg = result >> 31 != 0;
         let zero = result == 0;
         let overflow = ((lhs ^ result) & (rhs ^ result)) >> 31 != 0;
-        let carry = (result as u64) < (lhs as u64) + (rhs as u64);
+        let carry = u64::from(result) < u64::from(lhs) + u64::from(rhs);
         self.set_icc(neg, zero, overflow, carry);
     }
 
@@ -318,7 +318,7 @@ impl SparcState {
         let neg = result >> 31 != 0;
         let zero = result == 0;
         let overflow = ((lhs ^ rhs) & (lhs ^ result)) >> 31 != 0;
-        let carry = (lhs as u64) < (rhs as u64);
+        let carry = u64::from(lhs) < u64::from(rhs);
         self.set_icc(neg, zero, overflow, carry);
     }
 
@@ -365,7 +365,7 @@ impl SparcState {
         self.psr = (self.psr | PSR_S) & !PSR_ET;
         self.psr = (self.psr & !0x1F) | (old_cwp as u32 & 0x1F);
         // Build TBR with tt.
-        let tt_val = (tt.tt() as u32) << 4;
+        let tt_val = u32::from(tt.tt()) << 4;
         self.tbr = (self.tbr & !0xFF0) | tt_val;
         self.pc = self.tbr;
         self.npc = self.tbr + 4;
@@ -472,7 +472,7 @@ impl SparcState {
         // Sign-extend the 30-bit disp field to i32, then scale by 4.
         // We work in i64 to avoid any overflow before the sign extension,
         // sign-extend by shifting left then right, and truncate to i32.
-        let field = (instr & 0x3FFF_FFFF) as i64;
+        let field = i64::from(instr & 0x3FFF_FFFF);
         // Sign-extend 30-bit value: shift MSB to bit 63, arithmetic-shift back.
         let sign_ext = (field << 34) >> 34; // 64 - 30 = 34
         let disp30 = (sign_ext * 4) as i32;
@@ -630,22 +630,21 @@ impl SparcState {
                     .mem
                     .read_byte(addr)
                     .ok_or(SparcError::DataFault { addr, write: false })?;
-                self.rset(rd, val as u32);
+                self.rset(rd, u32::from(val));
             }
             0x02 => {
                 // LDUH (unsigned half)
                 // Use wrapping_add to avoid overflow when addr == 0xFFFF_FFFF
                 // (parser-trailing-data-ignored / integer-overflow guard).
-                let hi = self
+                let hi = u32::from(self
                     .mem
                     .read_byte(addr)
-                    .ok_or(SparcError::DataFault { addr, write: false })?
-                    as u32;
+                    .ok_or(SparcError::DataFault { addr, write: false })?);
                 let addr1 = addr.wrapping_add(1);
-                let lo = self.mem.read_byte(addr1).ok_or(SparcError::DataFault {
+                let lo = u32::from(self.mem.read_byte(addr1).ok_or(SparcError::DataFault {
                     addr: addr1,
                     write: false,
-                })? as u32;
+                })?);
                 self.rset(rd, (hi << 8) | lo);
             }
             0x09 => {
@@ -655,7 +654,7 @@ impl SparcState {
                     .read_byte(addr)
                     .ok_or(SparcError::DataFault { addr, write: false })?
                     as i8;
-                self.rset(rd, val as i32 as u32);
+                self.rset(rd, i32::from(val) as u32);
             }
             0x04 => {
                 // ST (32-bit store)
